+++ src/cli/actions/clear.rs
@@ -27,6 +27,7 @@ impl Runner {
                 "--symlinks",
                 "-s",
                 "--no-color",
+                "--debug",
             ],
         )?;