+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented here
 
 - added `backtrace` for errors
 - adding debug options to `inout`
+- added `--debug` flag to show debug output
 
 ## v0.11.0